+++ check.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
-./vendor/bin/phpmd . text phpmd.xml  --exclude vendor,updates,classes/seeders,tests
+./vendor/bin/phpmd . text phpmd.xml  --exclude vendor,updates,classes/seeders,tests,console
 ./vendor/bin/phpcbf
